+++ packages/edition-slices/src/tiles/tile-as/styles/index.js
@@ -1,0 +1,20 @@
+import { fonts, spacing } from "@times-components/styleguide";
+
+const styles = {
+  container: {
+    flex: 1,
+    paddingVertical: spacing(2),
+    paddingHorizontal: spacing(2)
+  },
+  headline: {
+    fontFamily: fonts.headline,
+    fontSize: 18,
+    lineHeight: 18
+  },
+  imageContainer: {
+    marginBottom: spacing(2),
+    width: "100%"
+  }
+};
+
+export default styles;