+++ packages/styleguide/__tests__/shared.js
@@ -3,6 +3,7 @@ import React from "react";
 import TestRenderer from "react-test-renderer";
 import { iterator } from "@times-components/test-utils";
 import styleguide, { Animations, scales } from "../src/styleguide";
+import themeFactory from "../src/theme/theme-factory";
 
 const tests = [
   {
@@ -73,7 +74,7 @@ const tests = [
   },
   {
     name:
-      "Line height factory should throw a a TypeError if no font or font size are provided",
+      "Line height factory should throw a TypeError if no font or font size are provided",
     test: () => {
       expect(() => styleguide().lineHeight()).toThrow(TypeError);
     }
@@ -88,6 +89,116 @@ const tests = [
           fontSize: "massive"
         })
       ).toThrow(TypeError);
+    }
+  },
+  {
+    name: "Theme Factory - DropCapFont should return the correct font",
+    test: () => {
+      expect(themeFactory("default", "indepth").dropCapFont).toEqual("dropCap");
+      expect(themeFactory("default", "maincomment").dropCapFont).toEqual(
+        "dropCap"
+      );
+      expect(themeFactory("culture", "indepth").dropCapFont).toEqual(
+        "cultureMagazine"
+      );
+      expect(themeFactory("style", "indepth").dropCapFont).toEqual(
+        "styleMagazine"
+      );
+      expect(
+        themeFactory("thesundaytimesmagazine", "magazinestandard").dropCapFont
+      ).toEqual("stMagazine");
+      expect(
+        themeFactory("thesundaytimesmagazine", "mainstandard").dropCapFont
+      ).toEqual("dropCap");
+      expect(themeFactory().dropCapFont).toEqual("dropCap");
+    }
+  },
+  {
+    name: "Theme Factory - HeadlineFont should return the correct font",
+    test: () => {
+      expect(themeFactory("default", "indepth").headlineFont).toEqual(
+        "headline"
+      );
+      expect(themeFactory("default", "maincomment").headlineFont).toEqual(
+        "headline"
+      );
+      expect(themeFactory("culture", "indepth").headlineFont).toEqual(
+        "cultureMagazine"
+      );
+      expect(themeFactory("style", "indepth").headlineFont).toEqual(
+        "styleMagazine"
+      );
+      expect(
+        themeFactory("thesundaytimesmagazine", "magazinestandard").headlineFont
+      ).toEqual("stMagazine");
+      expect(
+        themeFactory("thesundaytimesmagazine", "mainstandard").headlineFont
+      ).toEqual("headline");
+    }
+  },
+  {
+    name:
+      "Theme Factory - imageCaptionAlignment should return the correct caption alignment",
+    test: () => {
+      expect(themeFactory("", "indepth").imageCaptionAlignment).toStrictEqual({
+        primary: "center"
+      });
+      expect(
+        themeFactory("", "maincomment").imageCaptionAlignment
+      ).toStrictEqual({});
+    }
+  },
+  {
+    name: "Theme Factory - pullQuoteFont should return the correct font",
+    test: () => {
+      expect(themeFactory("default", "indepth").pullQuoteFont).toEqual(
+        "headlineRegular"
+      );
+      expect(themeFactory("default", "maincomment").pullQuoteFont).toEqual(
+        "headlineRegular"
+      );
+      expect(themeFactory("culture", "indepth").pullQuoteFont).toEqual(
+        "cultureMagazine"
+      );
+      expect(themeFactory("style", "indepth").pullQuoteFont).toEqual(
+        "styleMagazine"
+      );
+      expect(
+        themeFactory("thesundaytimesmagazine", "magazinestandard").pullQuoteFont
+      ).toEqual("stMagazine");
+      expect(
+        themeFactory("thesundaytimesmagazine", "mainstandard").pullQuoteFont
+      ).toEqual("headlineRegular");
+    }
+  },
+  {
+    name:
+      "Theme Factory - sectionColour should return the correct section with colours",
+    test: () => {
+      expect(themeFactory("default", "indepth").sectionColour).toEqual(
+        undefined
+      );
+      expect(themeFactory("default", "maincomment").sectionColour).toEqual(
+        "#1D1D1B"
+      );
+      expect(themeFactory("news", "mainstandard").sectionColour).toEqual(
+        "#13354E"
+      );
+      expect(
+        themeFactory("business", "magazinestandard").sectionColour
+      ).toEqual("#005B8D");
+      expect(themeFactory().sectionColour).toEqual("#1D1D1B");
+    }
+  },
+  {
+    name:
+      "Theme Factory - headlineCase should return the correct headline case",
+    test: () => {
+      expect(themeFactory("default", "indepth").headlineCase).toEqual(null);
+      expect(themeFactory("style", "indepth").headlineCase).toEqual(
+        "uppercase"
+      );
+      expect(themeFactory("style", "mainstandard").headlineCase).toEqual(null);
     }
   }
 ];