+++ packages/article-image/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.20](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.19...@times-components/article-image@7.3.20) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
+## [7.3.19](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.18...@times-components/article-image@7.3.19) (2019-08-06)
+
+
+### Bug Fixes
+
+* revert design changes for web ([#2161](https://github.com/newsuk/times-components/issues/2161)) ([667427f](https://github.com/newsuk/times-components/commit/667427f))
+
+
+
+
+
+## [7.3.18](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.17...@times-components/article-image@7.3.18) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/article-image
+
+
+
+
+
 ## [7.3.17](https://github.com/newsuk/times-components/compare/@times-components/article-image@7.3.16...@times-components/article-image@7.3.17) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/article-image