+++ packages/mock-tpa-server/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.19](https://github.com/newsuk/times-components/compare/@times-components/mock-tpa-server@0.2.18...@times-components/mock-tpa-server@0.2.19) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/mock-tpa-server
+
+
+
+
+
 ## [0.2.18](https://github.com/newsuk/times-components/compare/@times-components/mock-tpa-server@0.2.17...@times-components/mock-tpa-server@0.2.18) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/mock-tpa-server