+++ packages/provider/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.30](https://github.com/newsuk/times-components/compare/@times-components/provider@1.17.29...@times-components/provider@1.17.30) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/provider
+
+
+
+
+
+## [1.17.29](https://github.com/newsuk/times-components/compare/@times-components/provider@1.17.28...@times-components/provider@1.17.29) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/provider
+
+
+
+
+
 ## [1.17.28](https://github.com/newsuk/times-components/compare/@times-components/provider@1.17.27...@times-components/provider@1.17.28) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/provider