+++ packages/save-and-share-bar/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.11](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.7.10...@times-components/save-and-share-bar@0.7.11) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
+## [0.7.10](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.7.9...@times-components/save-and-share-bar@0.7.10) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
+## [0.7.9](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.7.8...@times-components/save-and-share-bar@0.7.9) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
 ## [0.7.8](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.7.7...@times-components/save-and-share-bar@0.7.8) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/save-and-share-bar