+++ packages/section/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.39](https://github.com/newsuk/times-components/compare/@times-components/section@1.1.38...@times-components/section@1.1.39) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.1.38](https://github.com/newsuk/times-components/compare/@times-components/section@1.1.37...@times-components/section@1.1.38) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
+## [1.1.37](https://github.com/newsuk/times-components/compare/@times-components/section@1.1.36...@times-components/section@1.1.37) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
 ## [1.1.36](https://github.com/newsuk/times-components/compare/@times-components/section@1.1.35...@times-components/section@1.1.36) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/section