+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.95](https://github.com/newsuk/times-components/compare/@times-components/topic@5.0.94...@times-components/topic@5.0.95) (2019-08-07)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.0.94](https://github.com/newsuk/times-components/compare/@times-components/topic@5.0.93...@times-components/topic@5.0.94) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.0.93](https://github.com/newsuk/times-components/compare/@times-components/topic@5.0.92...@times-components/topic@5.0.93) (2019-08-06)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 ## [5.0.92](https://github.com/newsuk/times-components/compare/@times-components/topic@5.0.91...@times-components/topic@5.0.92) (2019-08-05)
 
 **Note:** Version bump only for package @times-components/topic